requirements.txt
assemblyai
google-ai-generativelanguage
google-generativeai
gTTS
huggingface-hub
langchain
langchain-community
langchain-core
langchain-google-genai
numpy
opencv-python
pandas
pillow
plotly
pydantic
python-dotenv
requests
requests-oauthlib
requests-toolbelt
SpeechRecognition
streamlit

